dependency_graph.py: start to support "weak" labels
The `find_all_data_paths` function is useful for a script to use any
data that might be available from previous runs, including previous runs
of the same script -- and I have used it for exactly that purpose. This
"load any data that may be available" functionality can easily introduce
a cycle into the dependency graph, so track these edges separately. This
will shortly be used to drop the "weak" references from the graph before
performing a topological sort, to remove any such cycles.